chore: extract infra and encoding methods
