+++ lib/web/fetch/formdata-parser.js
@@ -1,11 +1,12 @@
 'use strict'
 
 const { bufferToLowerCasedHeaderName } = require('../../core/util')
-const { utf8DecodeBytes } = require('./util')
-const { HTTP_TOKEN_CODEPOINTS, isomorphicDecode } = require('./data-url')
+const { HTTP_TOKEN_CODEPOINTS } = require('./data-url')
 const { makeEntry } = require('./formdata')
 const { webidl } = require('../webidl')
 const assert = require('node:assert')
+const { isomorphicDecode } = require('../infra')
+const { utf8DecodeBytes } = require('../../encoding')
 
 const formDataNameBuffer = Buffer.from('form-data; name="')
 const filenameBuffer = Buffer.from('filename')
