+++ lib/web/fetch/index.js
@@ -35,7 +35,6 @@ const {
   isErrorLike,
   fullyReadBody,
   readableStreamClose,
-  isomorphicEncode,
   urlIsLocal,
   urlIsHttpHttpsScheme,
   urlHasHttpsScheme,
@@ -63,6 +62,7 @@ const { webidl } = require('../webidl')
 const { STATUS_CODES } = require('node:http')
 const { bytesMatch } = require('../subresource-integrity/subresource-integrity')
 const { createDeferredPromise } = require('../../util/promise')
+const { isomorphicEncode } = require('../infra')
 
 const hasZstd = typeof zlib.createZstdDecompress === 'function'
